Cleanup some stuff
