+++ api/src/main/kotlin/me/snoty/backend/scheduling/Scheduler.kt
@@ -1,5 +1,7 @@
 package me.snoty.backend.scheduling
 
+import org.jobrunr.jobs.lambdas.JobRequest
+
 interface Scheduler {
 	fun scheduleJob(id: String, job: () -> Unit)
