+++ integrations/moodle/src/main/kotlin/me/snoty/integration/moodle/model/MoodleAssignement.kt
@@ -8,7 +8,7 @@ import kotlinx.serialization.json.buildJsonObject
 import kotlinx.serialization.json.put
 import me.snoty.backend.integration.common.diff.UpdatableEntity
 import me.snoty.backend.integration.common.diff.Fields
-import me.snoty.backend.integration.moodle.model.raw.MoodleEvent
+import me.snoty.integration.moodle.model.raw.MoodleEvent
 import java.time.Instant
 import java.time.ZoneId
