+++ integrations/moodle/src/main/kotlin/me/snoty/integration/moodle/request/UserPreferencesRequest.kt
@@ -1,12 +1,8 @@
 package me.snoty.integration.moodle.request
 
-import me.snoty.backend.integration.moodle.*
-import me.snoty.backend.integration.moodle.model.MoodleUser
-import me.snoty.backend.integration.moodle.param.UserByFieldMoodleParam
-import me.snoty.integration.moodle.MoodleAPI
-import me.snoty.integration.moodle.MoodleRequest
-import me.snoty.integration.moodle.MoodleSettings
-import me.snoty.integration.moodle.request
+import me.snoty.integration.moodle.*
+import me.snoty.integration.moodle.model.MoodleUser
+import me.snoty.integration.moodle.param.UserByFieldMoodleParam
 
 suspend fun MoodleAPI.getUser(userSettings: MoodleSettings): MoodleUser? {
 	val request = MoodleRequest(userSettings) {
