--- src/main/kotlin/me/snoty/backend/scheduling/IntegrationFetchRequest.kt
@@ -1,9 +1,0 @@
-package me.snoty.backend.scheduling
-
-import org.jobrunr.jobs.lambdas.JobRequest
-
-class IntegrationFetchRequest : JobRequest {
-	override fun getJobRequestHandler(): Class<IntegrationFetchRequestHandler> {
-		return IntegrationFetchRequestHandler::class.java
-	}
-}
