--- src/main/kotlin/me/snoty/backend/scheduling/IntegrationFetchRequestHandler.kt
@@ -1,9 +1,0 @@
-package me.snoty.backend.scheduling
-
-import org.jobrunr.jobs.lambdas.JobRequestHandler
-
-class IntegrationFetchRequestHandler : JobRequestHandler<IntegrationFetchRequest> {
-	override fun run(jobRequest: IntegrationFetchRequest) {
-		TODO()
-	}
-}
